+++ cpp/include/dnv/vista/sdk/VIS.h
@@ -32,6 +32,7 @@
 
 #include "Codebooks.h"
 #include "Gmod.h"
+#include "GmodPath.h"
 #include "ImoNumber.h"
 #include "LocationBuilder.h"
 #include "ParsingErrors.h"